+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 1.3.1 (TBD)
 
 * geopackage-core version updated to 1.3.1
+* Bounding of degree projected boxes before Web Mercator transformations
 
 ## [1.3.0](https://github.com/ngageoint/geopackage-java/releases/tag/1.3.0) (06-27-2017)